requirements.txt
asgiref==3.4.1
Django==3.2.6
django-filter==2.4.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
Pillow==8.3.2
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.3.0
